# internal/app/docs/swagger.yaml
basePath: /
definitions:
  github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo:
    properties:
      limit:
        description: Límite de registros por página
        example: 10
        type: integer
      skip:
        description: Registros saltados
        example: 0
        type: integer
      total:
        description: Total de registros
        example: 100
        type: integer
      total_pages:
        description: Total de páginas
        example: 10
        type: integer
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.FieldError:
    properties:
      field:
        description: Nombre del campo con error
        example: email
        type: string
      message:
        description: Mensaje de error
        example: must be a valid email
        type: string
    type: object
  github_com_eren_dev_go_server_internal_shared_validation.ValidationError:
    properties:
      errors:
        description: Lista de errores por campo
        items:
          $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.FieldError'
        type: array
    type: object
...
  internal_modules_resources.CreateResourceDTO:
    properties:
      description:
        description: Descripción del recurso
        example: Módulo de citas veterinarias
        type: string
      name:
        description: 'Nombre del recurso (debe coincidir con el segmento de ruta,
          ej: appointments)'
        example: appointments
        minLength: 2
        type: string
      tenant_id:
        description: ID del tenant al que pertenece
        example: 507f1f77bcf86cd799439011
        type: string
    required:
    - name
    - tenant_id
    type: object
  internal_modules_resources.PaginatedResourcesResponse:
    properties:
      data:
        description: Lista de recursos
        items:
          $ref: '#/definitions/internal_modules_resources.ResourceResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo'
        description: Información de paginación
    type: object
  internal_modules_resources.ResourceResponse:
    properties:
      created_at:
        description: Fecha de creación
        type: string
      description:
        description: Descripción del recurso
        example: Módulo de citas veterinarias
        type: string
      id:
        description: ID único del recurso
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: Nombre del recurso
        example: appointments
        type: string
      tenant_id:
        description: ID del tenant
        example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        description: Fecha de actualización
        type: string
    type: object
  internal_modules_resources.UpdateResourceDTO:
    properties:
      description:
        description: Descripción del recurso
        example: Módulo de citas veterinarias
        type: string
      name:
        description: Nombre del recurso
        example: appointments
        type: string
    type: object
  internal_modules_roles.CreateRoleDTO:
    properties:
      description:
        description: Descripción del rol
        example: Rol para veterinarios
        type: string
      name:
        description: Nombre del rol
        example: veterinary
        minLength: 2
        type: string
      permissions_ids:
        description: IDs de permisos asignados (pares recurso+acción)
        example:
        - 507f1f77bcf86cd799439011
        items:
          type: string
        type: array
      resources_ids:
        description: IDs de recursos accesibles por este rol
        example:
        - 507f1f77bcf86cd799439011
        items:
          type: string
        type: array
      tenant_id:
        description: ID del tenant al que pertenece
        example: 507f1f77bcf86cd799439011
        type: string
    required:
    - name
    - tenant_id
    type: object
  internal_modules_roles.PaginatedRolesResponse:
    properties:
      data:
        description: Lista de roles
        items:
          $ref: '#/definitions/internal_modules_roles.RoleResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_pagination.PaginationInfo'
        description: Información de paginación
    type: object
  internal_modules_roles.PermissionRef:
    properties:
      action:
        description: Acción permitida
        example: get
        type: string
      id:
        description: ID del permiso
        example: 507f1f77bcf86cd799439011
        type: string
      resource:
        allOf:
        - $ref: '#/definitions/internal_modules_roles.ResourceRef'
        description: Recurso asociado
    type: object
  internal_modules_roles.ResourceRef:
    properties:
      description:
        description: Descripción
        example: Módulo de citas
        type: string
      id:
        description: ID del recurso
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: Nombre del recurso
        example: appointments
        type: string
    type: object
  internal_modules_roles.RoleResponse:
    properties:
      created_at:
        description: Fecha de creación
        type: string
      description:
        description: Descripción del rol
        example: Rol para veterinarios
        type: string
      id:
        description: ID único del rol
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: Nombre del rol
        example: veterinary
        type: string
      permissions:
        description: Permisos poblados (recurso + acción)
        items:
          $ref: '#/definitions/internal_modules_roles.PermissionRef'
        type: array
      resources:
        description: Recursos accesibles poblados
        items:
          $ref: '#/definitions/internal_modules_roles.ResourceRef'
        type: array
      tenant_id:
        description: ID del tenant
        example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        description: Fecha de actualización
        type: string
    type: object
  internal_modules_roles.UpdateRoleDTO:
    properties:
      description:
        description: Descripción del rol
        example: Rol para veterinarios
        type: string
      name:
        description: Nombre del rol
        example: veterinary
        type: string
      permissions_ids:
        description: IDs de permisos asignados
        items:
          type: string
        type: array
      resources_ids:
        description: IDs de recursos accesibles
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vetsify.com
    name: Vetsify Support
  description: |-
    API para plataforma veterinaria multi-tenant. Soporta panel de administración (staff) y app móvil (owners).

    ## Autenticación
    Todas las rutas protegidas requieren el header `Authorization: Bearer <token>`.
    - **Staff**: Obtén tokens via `POST /api/auth/login`
    - **Owners (mobile)**: Obtén tokens via `POST /mobile/auth/login`

    ## Multi-tenancy
    Las rutas que operan sobre datos de un tenant específico (patients, species, etc.) requieren el header `X-Tenant-ID` con el ObjectID del tenant.
    Este header debe enviarse en **todas** las peticiones a rutas marcadas con el parámetro `X-Tenant-ID`.
    El frontend puede configurarlo una vez en el interceptor HTTP (axios/fetch).

    ## Roles (RBAC)
    Las rutas del panel admin (`/api/*`) están protegidas por RBAC.
    Roles disponibles: `admin`, `veterinarian`, `receptionist`, `assistant`, `accountant`.

    ## Paginación
    Las rutas que retornan listas soportan `skip` y `limit` como query params.
    Respuesta incluye `pagination: { skip, limit, total, total_pages }`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Vetsify API
  version: "1.0"
paths:
...
  /api/resources:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de recursos
      parameters:
      - default: 0
        description: ' '
        in: query
        name: skip
        type: integer
      - default: 10
        description: ' '
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_resources.PaginatedResourcesResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar recursos
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Crea un nuevo recurso del sistema (debe coincidir con el segmento
        de ruta)
      parameters:
      - description: Datos del recurso
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_resources.CreateResourceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_resources.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Nombre ya existe
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear recurso
      tags:
      - resources
  /api/resources/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un recurso por su ID (soft delete)
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar recurso
      tags:
      - resources
    get:
      consumes:
      - application/json
      description: Obtiene un recurso por su ID
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_resources.ResourceResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener recurso
      tags:
      - resources
    patch:
      consumes:
      - application/json
      description: Actualiza un recurso existente
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_resources.UpdateResourceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_resources.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar recurso
      tags:
      - resources
  /api/roles:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de roles con permisos y recursos poblados
      parameters:
      - default: 0
        description: ' '
        in: query
        name: skip
        type: integer
      - default: 10
        description: ' '
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_roles.PaginatedRolesResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Crea un nuevo rol con permisos y recursos asignados
      parameters:
      - description: Datos del rol
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_roles.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_roles.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Nombre ya existe
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear rol
      tags:
      - roles
  /api/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un rol por su ID (soft delete)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar rol
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Obtiene un rol por su ID con permisos y recursos poblados
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_roles.RoleResponse'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener rol
      tags:
      - roles
    patch:
      consumes:
      - application/json
      description: Actualiza un rol, sus permisos y/o recursos asignados
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_roles.UpdateRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_roles.RoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eren_dev_go_server_internal_shared_validation.ValidationError'
        "401":
          description: No autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar rol
      tags:
      - roles
...
securityDefinitions:
  Bearer:
    description: 'Ingresa el token con el prefijo Bearer: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"